+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/operations/GetWithMetadataOperation.java
@@ -13,6 +13,7 @@ import org.infinispan.client.hotrod.impl.transport.TransportFactory;
 import org.infinispan.client.hotrod.logging.Log;
 import org.infinispan.client.hotrod.logging.LogFactory;
 
+
 /**
  * Corresponds to getWithMetadata operation as described by
  * <a href="http://community.jboss.org/wiki/HotRodProtocol">Hot Rod protocol specification</a>.
@@ -53,7 +54,7 @@ public class GetWithMetadataOperation<V> extends AbstractKeyOperation<MetadataVa
          }
          long version = transport.readLong();
          if (trace) {
-            log.tracef("Received version: %d", version);
+            log.tracef("Received version: %d", (Long) version);
          }
          V value = codec.readUnmarshallByteArray(transport, status);
          result = new MetadataValueImpl<V>(creation, lifespan, lastUsed, maxIdle, version, value);
