+++ core/src/main/java/org/infinispan/executors/SemaphoreCompletionService.java
@@ -86,7 +86,7 @@ public class SemaphoreCompletionService<T> implements CompletionService<T> {
    public Future<T> submit(final Callable<T> task) {
       QueueingTask futureTask = new QueueingTask(task);
       queue.add(futureTask);
-      if (trace) log.tracef("New task submitted, tasks in queue %d, available permits %d", queue.size(), semaphore.availablePermits());
+      if (trace) log.tracef("New task submitted, tasks in queue %d, available permits %d", (Integer)queue.size(), (Integer)semaphore.availablePermits());
       executeFront();
       return futureTask;
    }
@@ -95,7 +95,7 @@ public class SemaphoreCompletionService<T> implements CompletionService<T> {
    public Future<T> submit(final Runnable task, T result) {
       QueueingTask futureTask = new QueueingTask(task, result);
       queue.add(futureTask);
-      if (trace) log.tracef("New task submitted, tasks in queue %d, available permits %d", queue.size(), semaphore.availablePermits());
+      if (trace) log.tracef("New task submitted, tasks in queue %d, available permits %d", (Integer)queue.size(), (Integer)semaphore.availablePermits());
       executeFront();
       return futureTask;
    }
@@ -168,11 +168,11 @@ public class SemaphoreCompletionService<T> implements CompletionService<T> {
 
       private void runInternal() {
          try {
-            if (trace) log.tracef("Task started, tasks in queue %d, available permits %d", queue.size(), semaphore.availablePermits());
+            if (trace) log.tracef("Task started, tasks in queue %d, available permits %d", (Integer)queue.size(), (Integer)semaphore.availablePermits());
             super.run();
          } finally {
             completionQueue.offer(this);
-            if (trace) log.tracef("Task finished, tasks in queue %d, available permits %d", queue.size(), semaphore.availablePermits());
+            if (trace) log.tracef("Task finished, tasks in queue %d, available permits %d", (Integer)queue.size(), (Integer)semaphore.availablePermits());
          }
       }
    }
