+++ core/src/main/java/org/infinispan/interceptors/impl/SequentialInterceptorChainImpl.java
@@ -65,7 +65,7 @@ public class SequentialInterceptorChainImpl implements SequentialInterceptorChai
    @Start
    private void printChainInfo() {
       if (log.isDebugEnabled()) {
-         log.debugf("Interceptor chain size: %d", size());
+         log.debugf("Interceptor chain size: %d", (Integer)size());
          log.debugf("Interceptor chain is: %s", toString());
       }
    }
