+++ core/src/main/java/org/infinispan/stream/impl/AbstractCacheStream.java
@@ -388,7 +388,7 @@ public abstract class AbstractCacheStream<T, S extends BaseStream<T, S>, S2 exte
                Collection<CacheEntry<Object, Object>> localValue = op.performOperationRehashAware(results);
                // TODO: we can do this more efficiently - this hampers performance during rehash
                if (dm.getReadConsistentHash().equals(ch)) {
-                  log.tracef("Found local values %s for id %s", localValue.size(), id);
+                  log.tracef("Found local values %s for id %s", (Integer)localValue.size(), id);
                   results.onCompletion(null, segments, localValue);
                } else {
                   Set<Integer> ourSegments = ch.getPrimarySegmentsForOwner(localAddress);
